+++ instrumentation/opentelemetry-instrumentation-grpc/tests/test_server_interceptor.py
@@ -20,10 +20,10 @@ from concurrent import futures
 
 import grpc
 
-import opentelemetry.ext.grpc
+import opentelemetry.instrumentation.grpc
 from opentelemetry import trace
-from opentelemetry.ext.grpc import GrpcInstrumentorServer, server_interceptor
-from opentelemetry.ext.grpc.grpcext import intercept_server
+from opentelemetry.instrumentation.grpc import GrpcInstrumentorServer, server_interceptor
+from opentelemetry.instrumentation.grpc.grpcext import intercept_server
 from opentelemetry.sdk import trace as trace_sdk
 from opentelemetry.test.test_base import TestBase
 
@@ -76,7 +76,7 @@ class TestOpenTelemetryServerInterceptor(TestBase):
         span = spans_list[0]
         self.assertEqual(span.name, "test")
         self.assertIs(span.kind, trace.SpanKind.SERVER)
-        self.check_span_instrumentation_info(span, opentelemetry.ext.grpc)
+        self.check_span_instrumentation_info(span, opentelemetry.instrumentation.grpc)
         grpc_server_instrumentor.uninstrument()
 
     def test_uninstrument(self):
@@ -141,7 +141,7 @@ class TestOpenTelemetryServerInterceptor(TestBase):
         self.assertIs(span.kind, trace.SpanKind.SERVER)
 
         # Check version and name in span's instrumentation info
-        self.check_span_instrumentation_info(span, opentelemetry.ext.grpc)
+        self.check_span_instrumentation_info(span, opentelemetry.instrumentation.grpc)
 
     def test_span_lifetime(self):
         """Check that the span is active for the duration of the call."""